refactor: 바텀 바 버튼 선언 위치 변경
