+++ app/src/main/java/com/example/swap_it/MainActivity.kt
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.example.swap_it.ui.product_list.ProductListScreen
 import com.example.swap_it.ui.theme.SwapitTheme
 
 class MainActivity : ComponentActivity() {
@@ -20,7 +19,7 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    ProductListScreen().ProductScreen()
+                    MenuScreen().MenuScreen()
                 }
             }
         }
